requirements.txt
aiofiles==24.1.0
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
django-environ==0.11.2
psycopg2==2.9.9
